+++ src/main/java/com/skyegibney/finar/game/events/MoveMadeEvent.java
@@ -1,3 +1,5 @@
 package com.skyegibney.finar.game.events;
 
-public record MoveMadeEvent(int gameId, String player, int n, long player1Time, long player2Time) {}
+import java.util.UUID;
+
+public record MoveMadeEvent(long gameId, String player, int n, long player1Time, long player2Time) {}
