+++ src/main/java/com/skyegibney/finar/websockets/ClientMessage.java
@@ -1,3 +1,5 @@
 package com.skyegibney.finar.websockets;
 
-public record ClientMessage(int gameId, int lobbyId, String type, Object data) {}
+import java.util.UUID;
+
+public record ClientMessage(long gameId, int lobbyId, String type, Object data) {}
